Revert "Rename monitoringStatusValueFormatters to monitoringItemsValueFormatters"
This reverts commit 82d3a3f10d28583ff7f9e0ef8a5043523b34a766.
